chemical_visualizer/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.3
openpyxl==3.10.10
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.7
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
waitress==3.0.0
